+++ Sources/MapboxSearch/PublicAPI/Offline/SearchOfflineManager.swift
@@ -41,25 +41,52 @@ public class SearchOfflineManager {
         engine.setTileStore(searchTileStore.commonTileStore, completion: completion)
     }
 
-    /// Creates TilesetDescriptor for offline search index data with provided dataset name and version.
+    // MARK: - Tileset with name, version, and language parameters
+
+    /// Creates TilesetDescriptor for offline search index data with provided dataset name, version, and language.
+    /// Providing nil or excluding the language parameter will use the dataset name as-is.
+    /// Providing a language will append it to the name.
     /// - Parameters:
     ///   - dataset: dataset name
     ///   - version: dataset version
+    ///   - language: Provide a ISO 639-1 Code language from NSLocale. Values will be appended to the place dataset
+    /// name.
     /// - Returns: TilesetDescriptor for TileStore
-    public static func createTilesetDescriptor(dataset: String, version: String? = nil) -> MapboxCommon
+    public static func createTilesetDescriptor(
+        dataset: String,
+        version: String? = nil,
+        language: String? = nil
+    ) -> MapboxCommon
     .TilesetDescriptor {
-        CoreSearchEngineStatics.createTilesetDescriptor(dataset: dataset, version: version ?? "")
+        CoreSearchEngineStatics.createTilesetDescriptor(
+            dataset: dataset,
+            version: version ?? "",
+            language: language
+        )
     }
 
     /// Creates TilesetDescriptor for offline search boundaries with provided dataset name and version.
+    /// Providing nil or excluding the language parameter will use the places dataset name as-is.
+    /// Providing a language will append it to the name.
     /// - Parameters:
     ///   - dataset: dataset name
     ///   - version: dataset version
+    ///   - language: Provide a ISO 639-1 Code language from NSLocale. Values will be appended to the dataset name.
     /// - Returns: TilesetDescriptor for TileStore
-    public static func createPlacesTilesetDescriptor(dataset: String, version: String? = nil) -> MapboxCommon
+    public static func createPlacesTilesetDescriptor(
+        dataset: String,
+        version: String? = nil,
+        language: String? = nil
+    ) -> MapboxCommon
     .TilesetDescriptor {
-        CoreSearchEngineStatics.createPlacesTilesetDescriptor(dataset: dataset, version: version ?? "")
+        CoreSearchEngineStatics.createPlacesTilesetDescriptor(
+            dataset: dataset,
+            version: version ?? "",
+            language: language
+        )
     }
+
+    // MARK: - Default tileset
 
     /// Creates TilesetDescriptor for offline search index data using default dataset name.
     /// - Returns: TilesetDescriptor for TileStore